Rename to TableRepairJob
Rename ScheduledRepairJob to TableRepairJob
