+++ standalone-integration/src/test/java/com/ericsson/bss/cassandra/ecchronos/standalone/ITTableRepairJob.java
@@ -12,12 +12,46 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ericsson.bss.cassandra.ecchronos.osgi;
+package com.ericsson.bss.cassandra.ecchronos.standalone;
 
-import static org.awaitility.Awaitility.await;
-import static org.junit.Assert.assertTrue;
+import com.datastax.driver.core.Cluster;
+import com.datastax.driver.core.DataType;
+import com.datastax.driver.core.Host;
+import com.datastax.driver.core.Metadata;
+import com.datastax.driver.core.Session;
+import com.datastax.driver.core.TableMetadata;
+import com.datastax.driver.core.TokenRange;
+import com.datastax.driver.core.querybuilder.Insert;
+import com.datastax.driver.core.querybuilder.QueryBuilder;
+import com.datastax.driver.core.schemabuilder.AbstractCreateStatement;
+import com.datastax.driver.core.schemabuilder.KeyspaceOptions;
+import com.datastax.driver.core.schemabuilder.SchemaBuilder;
+import com.datastax.driver.core.utils.UUIDs;
+import com.ericsson.bss.cassandra.ecchronos.core.CASLockFactory;
+import com.ericsson.bss.cassandra.ecchronos.core.HostStatesImpl;
+import com.ericsson.bss.cassandra.ecchronos.core.metrics.TableRepairMetrics;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.DefaultRepairConfigurationProvider;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairConfiguration;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairEntry;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairHistoryProviderImpl;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairStateFactoryImpl;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairStatus;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairSchedulerImpl;
+import com.ericsson.bss.cassandra.ecchronos.core.scheduling.ScheduleManagerImpl;
+import com.ericsson.bss.cassandra.ecchronos.core.utils.LongTokenRange;
+import com.ericsson.bss.cassandra.ecchronos.core.utils.ReplicatedTableProviderImpl;
+import com.ericsson.bss.cassandra.ecchronos.core.utils.TableReference;
+import com.ericsson.bss.cassandra.ecchronos.fm.RepairFaultReporter;
+import com.google.common.collect.Sets;
+import net.jcip.annotations.NotThreadSafe;
+import org.assertj.core.util.Lists;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
 
-import java.io.IOException;
 import java.net.InetAddress;
 import java.util.ArrayList;
 import java.util.Date;
@@ -31,78 +65,100 @@ import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
-import com.datastax.driver.core.Metadata;
-import com.datastax.driver.core.Session;
-import com.ericsson.bss.cassandra.ecchronos.connection.NativeConnectionProvider;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairHistoryProvider;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairHistoryProviderImpl;
-import com.google.common.collect.Lists;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
+import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.anyLong;
+import static org.mockito.Matchers.anyMapOf;
+import static org.mockito.Matchers.eq;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
-import com.datastax.driver.core.DataType;
-import com.datastax.driver.core.Host;
-import com.datastax.driver.core.TableMetadata;
-import com.datastax.driver.core.TokenRange;
-import com.datastax.driver.core.querybuilder.Insert;
-import com.datastax.driver.core.querybuilder.QueryBuilder;
-import com.datastax.driver.core.schemabuilder.AbstractCreateStatement;
-import com.datastax.driver.core.schemabuilder.KeyspaceOptions;
-import com.datastax.driver.core.schemabuilder.SchemaBuilder;
-import com.datastax.driver.core.utils.UUIDs;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairEntry;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairStatus;
-import com.ericsson.bss.cassandra.ecchronos.core.utils.LongTokenRange;
-import com.ericsson.bss.cassandra.ecchronos.core.utils.TableReference;
-import com.google.common.collect.Sets;
-
-import org.ops4j.pax.exam.Configuration;
-import org.ops4j.pax.exam.Option;
-import org.ops4j.pax.exam.OptionUtils;
-import org.ops4j.pax.exam.cm.ConfigurationAdminOptions;
-import org.ops4j.pax.exam.junit.PaxExam;
-import org.ops4j.pax.exam.spi.reactors.ExamReactorStrategy;
-import org.ops4j.pax.exam.spi.reactors.PerMethod;
-
-import javax.inject.Inject;
-
-@RunWith(PaxExam.class)
-@ExamReactorStrategy(PerMethod.class)
-public class ITScheduledRepairJob extends TestBase
+@RunWith(MockitoJUnitRunner.class)
+@NotThreadSafe
+public class ITTableRepairJob extends TestBase
 {
-    private static final String REPAIR_CONFIGURATION_PID = "com.ericsson.bss.cassandra.ecchronos.core.osgi.DefaultRepairConfigurationProviderComponent";
-    private static final String SCHEDULE_MANAGER_PID = "com.ericsson.bss.cassandra.ecchronos.core.osgi.ScheduleManagerService";
+    @Mock
+    private RepairFaultReporter myFaultReporter;
 
-    private static final String CONFIGURATION_REPAIR_INTERVAL_SECONDS = "repairIntervalSeconds";
-    private static final String CONFIGURATION_SCHEDULE_INTERVAL_IN_SECONDS = "scheduleIntervalInSeconds";
+    @Mock
+    private TableRepairMetrics myTableRepairMetrics;
 
-    @Inject
-    NativeConnectionProvider myNativeConnectionProvider;
+    private Metadata myMetadata;
 
     private Session mySession;
-    private Metadata myMetadata;
+
     private Host myLocalHost;
 
-    private RepairHistoryProvider myRepairHistoryProvider;
+    private HostStatesImpl myHostStates;
+
+    private RepairHistoryProviderImpl myRepairHistoryProvider;
+
+    private DefaultRepairConfigurationProvider myDefaultRepairConfigurationProvider;
+
+    private RepairSchedulerImpl myRepairSchedulerImpl;
+
+    private ScheduleManagerImpl myScheduleManagerImpl;
+
+    private CASLockFactory myLockFactory;
 
     private List<String> myCreatedKeyspaces = new ArrayList<>();
-
-    @Configuration
-    public Option[] configure() throws IOException
-    {
-        return OptionUtils.combine(basicOptions(), scheduleManagerOptions(), repairSchedulerOptions());
-    }
 
     @Before
     public void init()
     {
-        mySession = myNativeConnectionProvider.getSession();
-        myMetadata = mySession.getCluster().getMetadata();
-        myLocalHost = myNativeConnectionProvider.getLocalHost();
+        myLocalHost = getNativeConnectionProvider().getLocalHost();
+        mySession = getNativeConnectionProvider().getSession();
+        Cluster cluster = mySession.getCluster();
+        myMetadata = cluster.getMetadata();
 
+        myHostStates = HostStatesImpl.builder()
+                .withRefreshIntervalInMs(1000)
+                .withJmxProxyFactory(getJmxProxyFactory())
+                .build();
         myRepairHistoryProvider = new RepairHistoryProviderImpl(mySession, s -> s);
+        ReplicatedTableProviderImpl replicatedTableProvider = new ReplicatedTableProviderImpl(myLocalHost, myMetadata);
+
+        myLockFactory = CASLockFactory.builder()
+                .withNativeConnectionProvider(getNativeConnectionProvider())
+                .withHostStates(myHostStates)
+                .withStatementDecorator(s -> s)
+                .build();
+
+        myScheduleManagerImpl = ScheduleManagerImpl.builder()
+                .withLockFactory(myLockFactory)
+                .withRunInterval(100, TimeUnit.MILLISECONDS)
+                .build();
+
+        RepairStateFactoryImpl repairStateFactory = RepairStateFactoryImpl.builder()
+                .withHost(myLocalHost)
+                .withHostStates(HostStatesImpl.builder()
+                        .withJmxProxyFactory(getJmxProxyFactory())
+                        .build())
+                .withMetadata(myMetadata)
+                .withRepairHistoryProvider(myRepairHistoryProvider)
+                .withTableRepairMetrics(myTableRepairMetrics)
+                .build();
+
+        myRepairSchedulerImpl = RepairSchedulerImpl.builder()
+                .withJmxProxyFactory(getJmxProxyFactory())
+                .withTableRepairMetrics(myTableRepairMetrics)
+                .withFaultReporter(myFaultReporter)
+                .withScheduleManager(myScheduleManagerImpl)
+                .withRepairStateFactory(repairStateFactory)
+                .build();
+
+        RepairConfiguration repairConfiguration = RepairConfiguration.newBuilder()
+                .withRepairInterval(60, TimeUnit.MINUTES)
+                .build();
+
+        myDefaultRepairConfigurationProvider = DefaultRepairConfigurationProvider.newBuilder()
+                .withReplicatedTableProvider(replicatedTableProvider)
+                .withRepairScheduler(myRepairSchedulerImpl)
+                .withCluster(cluster)
+                .withDefaultRepairConfiguration(repairConfiguration)
+                .build();
     }
 
     @After
@@ -120,6 +176,11 @@ public class ITScheduledRepairJob extends TestBase
 
             mySession.execute(SchemaBuilder.dropKeyspace(keyspace).ifExists());
         }
+        myHostStates.close();
+        myDefaultRepairConfigurationProvider.close();
+        myRepairSchedulerImpl.close();
+        myScheduleManagerImpl.close();
+        myLockFactory.close();
     }
 
     /**
@@ -141,6 +202,7 @@ public class ITScheduledRepairJob extends TestBase
         await().pollInterval(1, TimeUnit.SECONDS).atMost(90, TimeUnit.SECONDS).until(() -> isRepairedSince(tableReference, startTime));
 
         verifyTableRepairedSince(tableReference, startTime);
+        verify(myFaultReporter, never()).raise(any(RepairFaultReporter.FaultCode.class), anyMapOf(String.class, Object.class));
     }
 
     /**
@@ -167,6 +229,7 @@ public class ITScheduledRepairJob extends TestBase
 
         verifyTableRepairedSince(tableReference, startTime);
         verifyTableRepairedSince(tableReference2, startTime);
+        verify(myFaultReporter, never()).raise(any(RepairFaultReporter.FaultCode.class), anyMapOf(String.class, Object.class));
     }
 
     /**
@@ -198,6 +261,7 @@ public class ITScheduledRepairJob extends TestBase
         await().pollInterval(1, TimeUnit.SECONDS).atMost(90, TimeUnit.SECONDS).until(() -> isRepairedSince(tableReference, startTime, expectedRepairedRanges));
 
         verifyTableRepairedSince(tableReference, expectedRepairedInterval, expectedRepairedRanges);
+        verify(myFaultReporter, never()).raise(any(RepairFaultReporter.FaultCode.class), anyMapOf(String.class, Object.class));
     }
 
     private void verifyTableRepairedSince(TableReference tableReference, long repairedSince)
@@ -208,7 +272,12 @@ public class ITScheduledRepairJob extends TestBase
     private void verifyTableRepairedSince(TableReference tableReference, long repairedSince, Set<LongTokenRange> expectedRepaired)
     {
         OptionalLong repairedAt = lastRepairedSince(tableReference, repairedSince);
-        assertTrue(repairedAt.isPresent());
+        assertThat(repairedAt.isPresent()).isTrue();
+
+        verify(myTableRepairMetrics).lastRepairedAt(tableReference, repairedAt.getAsLong());
+
+        int expectedTokenRanges = expectedRepaired.size();
+        verify(myTableRepairMetrics, times(expectedTokenRanges)).repairTiming(eq(tableReference), anyLong(), any(TimeUnit.class), eq(true));
     }
 
     private boolean isRepairedSince(TableReference tableReference, long repairedSince)
@@ -351,19 +420,5 @@ public class ITScheduledRepairJob extends TestBase
         long start = (long) range.getStart().getValue();
         long end = (long) range.getEnd().getValue();
         return new LongTokenRange(start, end);
-    }
-
-    private Option repairSchedulerOptions()
-    {
-        return ConfigurationAdminOptions.newConfiguration(REPAIR_CONFIGURATION_PID)
-                .put(CONFIGURATION_REPAIR_INTERVAL_SECONDS, TimeUnit.HOURS.toSeconds(1))
-                .asOption();
-    }
-
-    private Option scheduleManagerOptions()
-    {
-        return ConfigurationAdminOptions.newConfiguration(SCHEDULE_MANAGER_PID)
-                .put(CONFIGURATION_SCHEDULE_INTERVAL_IN_SECONDS, 1L)
-                .asOption();
     }
 }
